requirements.txt
pandas==2.0.3
requests==2.31.0
psycopg2-binary==2.9.6
boto3==1.28.57
python-dotenv==1.0.0 